RP_Novel_Service/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==3.2
django-cors-headers==4.3.1
django-filter==2.4.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
idna==3.8
mysqlclient==2.0.3
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
